# parking/api/swagger/parking.yaml
---
swagger: "2.0"
info:
  description: "Parking Net Go project | Parking svc"
  version: "0.3.0"
  title: "parkings.parking"
tags:
  - name: "parking"
    description: "Parking operations"
  - name: "instruments"
    description: "Inner operations"
schemes:
  #  - "https"
  - "http"
paths:
  /parking/{parking_id}:
    get:
      tags:
        - "parking"
      summary: "Get parking place info by id"
      operationId: "get_parking_by_id"
      produces:
        - "application/json"
      parameters:
        - name: "parking_id"
          in: "path"
          description: "ID of parking place to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ParkingPlace"
        404:
          description: "Parking place not found"
          schema:
            $ref: "#/definitions/Error"

    put:
      tags:
        - "parking"
      summary: "Update parking place"
      operationId: "update_parking"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "parking_id"
          in: "path"
          description: "ID of parking place to change"
          required: true
          type: "integer"
          format: "int64"
        - name: "object"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/ParkingPlace"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ParkingPlace"
        403:
          description: "No access"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "No such element"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Incorrect data"
          schema:
            $ref: "#/definitions/Error"
      security:
        - api_key: [ ]

    delete:
      tags:
        - "parking"
      summary: "Delete parking place"
      operationId: "delete_parking"
      produces:
        - "application/json"
      parameters:
        - name: "parking_id"
          in: "path"
          description: "ID of parking place to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Result"
        403:
          description: "No access"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Parking place not found"
          schema:
            $ref: "#/definitions/Error"
      security:
        - api_key: [ ]

  /parking:
    get:
      tags:
        - "parking"
      summary: "Get suitable parking places"
      operationId: "get_parkings"
      produces:
        - "application/json"
      parameters:
        - name: "city"
          in: "query"
          type: "string"
        - name: "name"
          in: "query"
          type: "string"
        - name: "parking_type"
          in: "query"
          type: "string"
          enum:
            - "outdoor"
            - "covered"
            - "underground"
            - "multi-level"
        - name: "owner_id"
          in: "query"
          type: "string"
          description: "Filter parking places by owner ID (for owners to get their own parkings)"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ParkingPlace"
        404:
          description: "Suitable parking places not found"
          schema:
            $ref: "#/definitions/Error"

    post:
      tags:
        - "parking"
      summary: "Create parking place"
      operationId: "create_parking"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "object"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/ParkingPlace"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
        403:
          description: "No access"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Incorrect data"
          schema:
            $ref: "#/definitions/Error"
      security:
        - api_key: [ ]
  /metrics:
    get:
      tags:
        - "instruments"
      summary: "Prometheus metrics"
      produces:
        - "application/json"
      responses:
        200:
          description: ok
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  ParkingPlace:
    type: "object"
    required:
      - "name"
      - "city"
      - "address"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        example: "Central Parking"
      city:
        type: "string"
        example: "Moscow"
      address:
        type: "string"
        example: "Red Square №1"
      parking_type:
        type: "string"
        description: "type of parking facility"
        enum:
          - "outdoor"
          - "covered"
          - "underground"
          - "multi-level"
      hourly_rate:
        type: "integer"
        format: "int64"
        description: "hourly parking rate"
      capacity:
        type: "integer"
        format: "int64"
        description: "total number of parking spots"
      owner_id:
        type: "string"
  Error:
    type: "object"
    required:
      - "error_status_code"
    properties:
      error_message:
        type: "string"
      error_status_code:
        type: integer
  Result:
    type: "object"
    properties:
      status:
        type: "string"
      message:
        type: "string"